chat_requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
websockets==11.0.3
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-jose[cryptography]